requirements.txt
# GX TradeIntel v6 — Dependencies
smartapi-python==1.5.5
kiteconnect>=5.0.0
pyotp==2.9.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
logzero>=1.7.0
pytz>=2023.3
yfinance>=0.2.40
